requirements.txt

SpeechRecognition
wordcloud 
matplotlib
pydub
google.generativeai
streamlit-audiorec
streamlit